chỉnh lại quy trinh order cho hop ly

+++ Fullstackecommercereactjswithadminsourcecode/admin/src/pages/BatchCodes/requestBatchCode.js
@@ -63,13 +63,31 @@ const RequestBatchCode = () => {
     const deleteBatchCode = (id) => {
         const confirmDelete = window.confirm("Are you sure you want to delete?");
         if (!confirmDelete) return;
-        setIsLoading(true);
-        context.setProgress(30);
-        deleteData(`/api/batchCodes/${id}`).then(() => {
-            setBatchCodes(batchCodes.filter(batch => batch._id !== id));
-            context.setProgress(100);
-            setIsLoading(false);
-        });
+        const userInfo = JSON.parse(localStorage.getItem("user"));
+        if (userInfo?.role === "mainAdmin") {
+            setIsLoading(true);
+            context.setProgress(30);
+            deleteData(`/api/batchCodes/${id}`).then(() => {
+                setBatchCodes(batchCodes.filter(batch => batch._id !== id));
+                context.setProgress(100);
+                context.setAlertBox({
+                    open: true,
+                    error: false,
+                    msg: "Batch Code Deleted!"
+                });
+                setIsLoading(false);
+                fetchDataFromApi(`/api/batchCodes/locationBatchCode?locationId=${user.locationId}&locationName=${user.locationName}&percentage=${percentRemain}&expiredDay=${dayRemain}&q=${querySearch}&page=1&perPage=${perPage}`).then((res) => {
+                    setBatchCodes(res);
+                    context.setProgress(100);
+                });
+            });
+        } else {
+            context.setAlertBox({
+                open: true,
+                error: true,
+                msg: "Only Admin can delete ",
+            });
+        }
     };
 
     const handleChangeStatus = (e, id) => {
