+++ Fullstackecommercereactjswithadminsourcecode/admin/src/pages/Users/index.js
@@ -61,14 +61,23 @@ const Users = () => {
   const deleteUser = (id) => {
     const confirmDelete = window.confirm("Are you sure you want to delete?");
     if (!confirmDelete) return;
-    deleteData(`/api/user/${id}`).then(() => {
-      context.setAlertBox({
-        open: true,
-        error: false,
-        msg: "User Deleted!",
+    const userInfo = JSON.parse(localStorage.getItem("user"));
+    if (userInfo?.role === "mainAdmin") {
+      deleteData(`/api/user/${id}`).then(() => {
+        context.setAlertBox({
+          open: true,
+          error: false,
+          msg: "User Deleted!",
+        });
+        loadUsers(page, perPage);
       });
-      loadUsers(page, perPage);
-    });
+    } else {
+        context.setAlertBox({
+            open: true,
+            error: true,
+            msg: "Only Admin can delete ",
+        });
+    }
   };
 
   const handleChange = (event, value) => {
